refactor: RacingManager 리펙토링

+++ src/main/java/racingcar/util/IntConverter.java
@@ -8,7 +8,7 @@ public class IntConverter {
     private IntConverter() {
     }
 
-    public static long StringToInt(String string) {
+    public static int StringToInt(String string) {
         int parsedLong;
 
         try {
